Thread safety for stable-pointers (#36 in cl-cffi-gtk)
;; git log --no-walk --format=%B ed429b8 (enometh/cl-cffi-gtk)
Thread safety for stable-pointers (#36)

 #+begin_original_commit ferada sharpshifter branch
 commit 303a23666ce898f3848de75b9352487f41b98763
 Author:     bpatrikm <34801205+bpatrikm@users.noreply.github.com>
 AuthorDate: Mon Mar 30 00:27:46 2020 +0200
 Commit:     GitHub <noreply@github.com>
 CommitDate: Mon Mar 30 00:27:46 2020 +0200
 #+end

* Wrapping stable-pointer accesses with SBCL mutex

This works for me, in the sense that see no errors where nil gets funcall:ed. But it's a very naive approach, as I don't have a lot of experience with multi-threading.

* Wrapping stable-pointer accesses in mutex-locks.

This seems to work for me, in the sense that I don't see errors where nil gets funcall:ed. But I don't know if it might create other problems. Bordeaux-threads has to be loaded.

* Changed name sp-mutex -> stable-pointers lock

* bordeaux-threads now used by stable-pointers